+++ src/store.js
@@ -1,7 +1,7 @@
 import { combineReducers, createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly';
-import { connectRouter, routerMiddleware } from 'connected-react-router';
+import { createReduxHistoryContext } from 'redux-first-history';
 import { save, load } from 'redux-localstorage-simple';
 
 import config from '@plone/volto/registry';
@@ -10,9 +10,17 @@ import reducers from '@root/reducers';
 import { api, crashReporter, blacklistRoutes } from '@plone/volto/middleware';
 
 const configureStore = (initialState, history, apiHelper) => {
+  const {
+    createReduxHistory,
+    routerMiddleware,
+    routerReducer,
+  } = createReduxHistoryContext({
+    history,
+  });
+
   let stack = [
     blacklistRoutes,
-    routerMiddleware(history),
+    routerMiddleware,
     crashReporter,
     thunk,
     ...(apiHelper ? [api(apiHelper)] : []),
@@ -27,7 +35,7 @@ const configureStore = (initialState, history, apiHelper) => {
   const middlewares = composeWithDevTools(applyMiddleware(...stack));
   const store = createStore(
     combineReducers({
-      router: connectRouter(history),
+      router: routerReducer,
       ...reducers,
       ...config.addonReducers,
     }),
@@ -40,7 +48,7 @@ const configureStore = (initialState, history, apiHelper) => {
     middlewares,
   );
 
-  return store;
+  return [store, createReduxHistory(store)];
 };
 
 export default configureStore;